Extract message formatting from UniversalSearchStore
Lets other stores receiving messages to format them in the same way.
